fix: extract shared LocationSection from ShopLocationSection
